[SDK] feat: Add editable token amount input to buy screen
