+++ .changeset/nine-seahorses-judge.md
@@ -1,0 +1,5 @@
+---
+"thirdweb": patch
+---
+
+Allow editing pay amount mid flow
